% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{addClone}
\alias{addClone}
\title{add a single CNV clone in the object}
\usage{
addClone(X, gr, clonename = NA, type = c("snv", "cnv"))
}
\arguments{
\item{X}{cloneobj}

\item{gr}{GRanges withe the segments of a specific clone. The column 'CCF' of this Granges object is used to get the CCF of this clone. Alternatively the column 'VAF' can be used only if CCF is not present}

\item{clonename}{the name of the clone that will be added in teh list of clones.}

\item{type}{either 'snv' or 'cnv' depending on the type of events that are added}
}
\value{
an updated cloneobj
}
\description{
add a single CNV clone in the object
}
