% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{addCloneEntry}
\alias{addCloneEntry}
\title{add an entry for a new clone in the clone_info data frame}
\usage{
addCloneEntry(X, name, fraction)
}
\arguments{
\item{X}{cloneobj}

\item{name}{the name of the clone}

\item{fraction}{the ccf of the clone}
}
\value{
an updated cloneobj
}
\description{
add an entry for a new clone in the clone_info data frame
}
