% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{addCNVClones}
\alias{addCNVClones}
\title{add clones to the object
it uses a GRangesList as input, which contains all teh CNV segments for each clone
it expects a column named 'cnv.ccf' which contains the CCF of the clone.
The overall CCF of this clone is the mean of this value.}
\usage{
addCNVClones(X, cnv, name = NULL)
}
\arguments{
\item{X}{cloneobj}

\item{cnv}{GRangesList or GRanges object. If a list is provided the names of the list should correspond to the names of the clones
if no names are provided then they are created followign the pattern "cloneX" where X is 1, 2, ..}

\item{names}{a vector with names that will be used to name the GRanges objects. Should have the same length as the list of GRanges.
This value overrides the names provided as names of the list.}
}
\description{
add clones to the object
it uses a GRangesList as input, which contains all teh CNV segments for each clone
it expects a column named 'cnv.ccf' which contains the CCF of the clone.
The overall CCF of this clone is the mean of this value.
}
