% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{addEvents}
\alias{addEvents}
\title{perform final tasks with an object}
\usage{
addEvents(X, events, type = c("snv", "cnv"))
}
\arguments{
\item{X}{cloneobj}

\item{events}{a GRanges object. If a list is provided the names of the list should correspond to the names of the clones
if no names are provided then they are created followign the pattern "cloneX" where X is 1, 2, ..}

\item{type}{either 'snv' or 'cnv' depending on the type of events that are added}
}
\description{
perform final tasks with an object
}
