% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{addSNVClone}
\alias{addSNVClone}
\title{add a single SNV clone in the object}
\usage{
addSNVClone(X, gr, clonename = NA)
}
\arguments{
\item{X}{cloneobj}

\item{gr}{GRanges withe the segments of a specific clone. The column 'CCF' of this Granges object is used to get the CCF of this clone.}

\item{clonename}{the name of the clone that will be added in teh list of clones.}
}
\description{
add a single SNV clone in the object
}
