% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneman-package.R
\docType{package}
\name{cloneman}
\alias{cloneman}
\alias{cloneman-package}
\title{cloneman: a package to parse and manage information from Copy Number calling and Clone Inference tools.}
\description{
There exist many programs that are used to infer the copy number abberations in samples
They typically infer the copy number segments (either as integer numbers, e.g. 1, 3, 4 copies etc)
or as log ratios to the "normal" coverage.
Each program contains produces its own specific output format, in one or multiple files.

This package was conceived in an attempt to normalize these outputs in a standard R format
with a set of predetermined columns and objects that can be used for downstream analysis.
}
\section{CNV segments}{


The simplest parsed output of these tools is a list of segments.
More specifically for CNVs a GRanges object is produced with all the segments and their
associated information

The following columns are added to the metadata columns of the GRanges

\strong{total.allele.copies}: the total number of alleles of that segment,

\strong{major.allele.copies}: the number of copies of the \emph{major} allele,

\strong{minor.allele.copies}: the number of copies of the \emph{minor} allele,

\strong{normal.allele.copies}: the number of copies expected for a normal sample (typically coming from the normal ploidy of the sample)

\strong{cnv.flag}: For each segment depending on the combination of the above values a flag is added to indicate the type and
value of the aberration according to the schema:\tabular{lll}{
   \strong{total.allele.copies} \tab \strong{minor.allele.copies} \tab \strong{cnv.flag} \cr
   0 \tab 0 \tab HOMDEL \cr
   0 \tab NA \tab HOMDEL \cr
   1 \tab 0 \tab HETDEL \cr
   1 \tab NA \tab HETDEL \cr
   2 \tab 0 \tab NEUT-LOH \cr
   2 \tab 1 \tab NEUT \cr
   2 \tab NA \tab NEUT/Unknown \cr
   2+ \tab 0 \tab AMP-LOH \cr
   2+ \tab 1+ \tab AMP \cr
   2+ \tab NA \tab AMP-\link{LOH?} \cr
}


For methods where the ploidy is predicted it is parsed and added in the column \emph{ploidy}.
Since this package was developed with human genome in mind a column caled \emph{normal.allele.copies} is added
with the default value of '2'
and for the Y (chrY) chromosome this is set to 1
}

\section{Clone Abundance}{


The clone abundance can be calculated by the CNV calling method, but can also be computed from SNVs, using tools such as Sciclone, Pyclone, Pyclone-vi, Fastclone etc.
There are two ways of expressing the clone abundance
one as \emph{VAF} (variant allele frequency) and as \emph{CCF} (cellular clone frequency).
Since we typically assume a diploid baseline, the VAF is typically the frequence of the mutation observed on a single allele and it is half of the CCF.

To avoid this confusion we include both in the results

\strong{variant.allele.frequency}: represents the VAF of an event. Unless there is a way to compute it from other data, it is defaulted to NA.

\strong{cellular.abundance}: represents the CCF of an event.

In all cases they are expressed in fractions that are ranging between 0 and 1.
}

\section{list of supported tools}{
\tabular{lll}{
   \strong{tool} \tab \strong{type of data} \tab \strong{comments} \cr
   \href{https://pubmed.ncbi.nlm.nih.gov/27270079/}{FACETS} \tab CNV/CI/purity/ploidy \tab  \cr
   \href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-016-1174-7}{CloneCNA} \tab CNV/CI/purity/ploidy \tab  \cr
   \href{http://bioinformatics.oxfordjournals.org/content/28/3/423.long}{ControlFreec} \tab CNV/CI/purity/ploidy \tab CI not working well, ignored in this version \cr
}
}

\section{cloneman functions}{

\strong{Simple parsers}

\code{\link[=parseFACETS]{parseFACETS()}} : parse a dataframe produced by FACETS.

\code{\link[=parseFACETS_rds]{parseFACETS_rds()}}: load an Rdata file produced by FACETS. This is the preferred method in order to retain additional information.

\code{\link[=parseFACETS_file]{parseFACETS_file()}}: load a tab delimited file with segments produced by FACETS. Additional information (e.g. purity) is not retained

\code{\link[=parseCloneCNA]{parseCloneCNA()}}: parse a dataframe produced by CloneCNA

\code{\link[=parseCloneCNA_file]{parseCloneCNA_file()}}: parse a file produced by CloneCNA. This is the preferred method in order to retain additional information.

\code{\link[=parseCfreec]{parseCfreec()}}: parse a dataframe with segments produced by ControlFreec.

\code{\link[=parseCfreec_file]{parseCfreec_file()}}: parse a _CNV or _CNV.pvalue.txt file produced by ControlFreec
}

\keyword{internal}
