% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{cloneobj}
\alias{cloneobj}
\title{create the object}
\usage{
cloneobj(
  purity = 1,
  ploidy = 2,
  cnvlist = NULL,
  snvlist = NULL,
  method = "Unspecified",
  sample_name = "sample",
  genome = "hg38"
)
}
\arguments{
\item{purity}{(float) Purity of the sample ,}

\item{ploidy}{(float) Ploidy of the sample,}

\item{cnvlist}{GRanges with cnv events,}

\item{snvlist}{GRanges with snv events,}

\item{method}{(list of string) Name of method used to predict these values}

\item{genome}{the name of the genome (hg19, hg38)}
}
\value{
an object of type cloneobj
}
\description{
cloneobj is a structure that holds information
for the clones
}
\details{
This information includes

\strong{Method name} the name(s) of the method(s) used to create the data for this object

\strong{Purity}: the predicted purity of the sample (0-1)

\strong{Ploidy}: the predicted ploidy of the sample (default 2)

\strong{Copy number events} GRanges with segments from one or multiple clones (depending on the tool used). Each segment with its own CCF (if available)

\strong{SNV clones} GRanges of SNV events, with the SNV VAF (or CCF).

\strong{Sample Sex} either provided by the user or inferred'
}
