% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCNV.R
\name{cnvFlag}
\alias{cnvFlag}
\title{get the flag for a CNV,
according to the table
|---------|----------|--------------|---------------------------------------------------------------------|
|total cn | minor cn |	type         | meaning of type                                                     |
|---------|----------|--------------|---------------------------------------------------------------------|
|0        |	0 	    | HOMDEL       | Homozygous deletion                                                 |
|0        |	NA 	    | HOMDEL       | Homozygous deletion                                                 |
|1        |	0 	    | HETDEL       | Heterozygous deletion (1 out of 2 alleles)                          |
|1        |	NA 	    | HETDEL       | Heterozygous deletion (this is an edge case that should not occur)  |
|2        |	0 	    | NEUT-LOH     | Loss of heterozygosity without change in the number of alleles      |
|2        |	1 	    | NEUT         | Normal                                                              |
|2        |	NA 	    | NEUT/Unknown | Normal or undetermined by the algorithm                             |
|2+       |	0 	    | AMP-LOH      | Amplification with loss of heterozygosity                           |
|2+       |	1+ 	    | AMP          | Amplification                                                       |
|2+       |	NA 	    | AMP-[LOH?]   | Amplification with (possible) loss of heterozygosity                |
|---------|----------|--------------|---------------------------------------------------------------------|}
\usage{
cnvFlag(major.cn = NA, minor.cn = 0, total.cn = NA)
}
\description{
get the flag for a CNV,
according to the table
|---------|----------|--------------|---------------------------------------------------------------------|
|total cn | minor cn |	type         | meaning of type                                                     |
|---------|----------|--------------|---------------------------------------------------------------------|
|0        |	0 	    | HOMDEL       | Homozygous deletion                                                 |
|0        |	NA 	    | HOMDEL       | Homozygous deletion                                                 |
|1        |	0 	    | HETDEL       | Heterozygous deletion (1 out of 2 alleles)                          |
|1        |	NA 	    | HETDEL       | Heterozygous deletion (this is an edge case that should not occur)  |
|2        |	0 	    | NEUT-LOH     | Loss of heterozygosity without change in the number of alleles      |
|2        |	1 	    | NEUT         | Normal                                                              |
|2        |	NA 	    | NEUT/Unknown | Normal or undetermined by the algorithm                             |
|2+       |	0 	    | AMP-LOH      | Amplification with loss of heterozygosity                           |
|2+       |	1+ 	    | AMP          | Amplification                                                       |
|2+       |	NA 	    | AMP-[LOH?]   | Amplification with (possible) loss of heterozygosity                |
|---------|----------|--------------|---------------------------------------------------------------------|
}
