% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareCNV}
\alias{compareCNV}
\title{Compare the CNVs between two or more cloneobj}
\usage{
compareCNV(query, subject, method = "status", wiggle = 0.1)
}
\arguments{
\item{query}{the query cloneobj}

\item{subject}{the subject cloneobj}

\item{method}{The method of comparison to employ (based on the description above). Options are 'status','cnv','ccf'}

\item{wiggle}{The differnce between the CCF when compared that is considered acceptable}

\item{plot}{if set to TRUE draw an alluvial plot that shows how events are distributed between samples.}
}
\value{
a table with the results
}
\description{
Comparing the CNV segments between two methods is trick
Segments with slightly different coordinates may be overlapping by most
of their length, but they are not identical
In the case of WES data, we can compare the CNV status of each of the targets
This assumes that the dataset has been filtered to select only those regions
}
\details{
This function can compare two cloneobj and find how many segments have
\enumerate{
\item the same status (regardless of copy number), e.g. two segments with flag AMP
will be considered the same even if one is 4 copies and the other 8
\item the same status and same copy number.
\item the same status and same CCF. In this case the CCF is normalized for purity
}
}
