% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareStatus}
\alias{compareStatus}
\title{Compare two cloneobj by status}
\usage{
compareStatus(query, subject, wiggle = 0.1)
}
\arguments{
\item{query}{cloneobj}

\item{subject}{cloneobj}
}
\value{
a dataframe with the status, cnv and ccf for each overlapping segment
}
\description{
We use the cnv.flag to do the comparisons
Importatn assumptions: both methods have the same segments
}
\details{
For the CNV comparison in mode status:
TP: same in query and subject
FP: NEUT in query and something else in subject
FN: something else in query NEUT in subject

For the CNV comparison in mode cnv (compare the actual segment copy number)
The comparison is restricted to the TP segments of the 'status' part
TP: same in query and subject
FP: query has less copies than subject (more copies in subject)
FN: query has more copies than subject (less copies in subject)

For the CNV comparison in mode ccf (compare the CCF of each segment)
The comparison is restricted to the TP segments of the 'status' part
TP: |query CCF - subject CCF| < wiggle
FP:
}
