% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{computeStatistics}
\alias{computeStatistics}
\title{Internal method to calculate statistic}
\usage{
computeStatistics(df, method = "status")
}
\arguments{
\item{the}{dataframe with the status information}
}
\value{
a data frame with statistics
}
\description{
Internal method to calculate statistic
}
