% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{.fixNames}
\alias{.fixNames}
\title{fix the names of the clones in the list.
if a name vector is provided use that as the names of the list
else use the names that are assigned to the objects already
and if no names are assigned use the format 'cloneX'}
\usage{
.fixNames(gr, clone_names = NULL)
}
\arguments{
\item{gr}{GRangesList or GRanges object that will be assigned the names}

\item{clone_names}{a vector with names to replace the names of the elements}
}
\value{
a vector with the names of the elements of the list
}
\description{
fix the names of the clones in the list.
if a name vector is provided use that as the names of the list
else use the names that are assigned to the objects already
and if no names are assigned use the format 'cloneX'
}
