% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.maxCNV}
\alias{.maxCNV}
\title{helper function to find the maximum number of CNV copies}
\usage{
.maxCNV(X)
}
\arguments{
\item{X}{cloneobj}
}
\value{
Returns the maximum copy number among all the clones.
}
\description{
helper function to find the maximum number of CNV copies
}
