% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfreq.R
\name{.parseSuperFreq_CNV}
\alias{.parseSuperFreq_CNV}
\title{Parse the CNV segments from SuperFreq}
\usage{
.parseSuperFreq_CNV(superfreq.gr, cnasegments.fn)
}
\arguments{
\item{superfreq.gr}{Grages with SuperFreq output (loaded by parseSuperFreq_River)}

\item{cnasegments.fn}{file with the details of the CNV prediction}

\item{clone_info}{dataframe with the information about the clones}
}
\value{
A GRanges object with the segments predicted by SuperFreq, with clonal ifnormation added
}
\description{
Generate a cnv Granges from a dataframe created by SuperFreq
it matches the CNVs in the suprefreq.gr to the CNVs listed in the CNA segments file
and transfers the information from the latter to the former
it also adds the clonal information of the clones
}
