% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfreq.R
\name{.parseSuperFreq_info}
\alias{.parseSuperFreq_info}
\title{Parse the info file that Superfreq produces}
\usage{
.parseSuperFreq_info(clone_info)
}
\arguments{
\item{clone_info}{the file with the clone info (data/clones_{samplename}.tsv)}
}
\value{
a dataframe with the clones information.
}
\description{
This a file following the pattern data/clones_{samplename}.tsv
}
