% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfreq.R
\name{.parseSuperFreq_River}
\alias{.parseSuperFreq_River}
\title{Parse the rivers/{samplename}-river.tsv file}
\usage{
.parseSuperFreq_River(subclone_events_info)
}
\arguments{
\item{fn}{the {samplename}-river.tsv file}
}
\value{
a GRanges object with the events.
}
\description{
which contains the summary of CNV and SNV and their assignment to clones
}
