% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfreq.R
\name{.parseSuperFreq_SNV}
\alias{.parseSuperFreq_SNV}
\title{Extract the SNV events from the clones from SuperFreq}
\usage{
.parseSuperFreq_SNV(superfreq.gr, variants.fn)
}
\arguments{
\item{superfreq.gr}{Grages with SuperFreq output (loaded by parseSuperFreq_River)}

\item{variants.fn}{file with the details of the CNV prediction}

\item{clone_info}{dataframe with the information about the clones}
}
\value{
A GRanges object with the SNVs predicted by SuperFreq, with clonal information added (somaticVariants.tsv)
}
\description{
Extract the SNV events from the clones from SuperFreq
}
