% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfreq.R
\name{.SuperFreqFilenames}
\alias{.SuperFreqFilenames}
\title{Search the directory of SuperFreq and get the list of files that are needed for hte analysis}
\usage{
.SuperFreqFilenames(method_dir)
}
\arguments{
\item{method_dir}{directory created by superfreq. Under it are expected the directories data and rivers}
}
\value{
a list with the filenames of the filetypes needed
}
\description{
Search the directory of SuperFreq and get the list of files that are needed for hte analysis
}
