% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{filter}
\alias{filter}
\title{Filter regions}
\usage{
filter(X)
}
\arguments{
\item{X}{cloneobj to filter}

\item{gr}{GRanges object with the coordinates to filter}
}
\value{
a new object with the filtered coordinates.
}
\description{
Provide a list of regions that will be used to fragment the CNV and SVN of the object
This is useful to compare objects between them
and restrict the results on specific regions of the genome (e.g. the coordinates of an exome kit)
}
