% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishplot.R
\name{fishPlot}
\alias{fishPlot}
\title{create a fishplot using the snv events (if available)}
\usage{
fishPlot(x)
}
\arguments{
\item{x}{a cloneobj}
}
\description{
create a fishplot using the snv events (if available)
}
