% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{getCombinedCNV}
\alias{getCombinedCNV}
\title{total CNV segments}
\usage{
getCombinedCNV(X)
}
\arguments{
\item{X}{cloneobj}
}
\value{
a Granges object with combined CNV segments
}
\description{
create a list of all the segments and their fraction
e.g. if a segment is in two clones, its fraction is the sum of the clones
at the end the result contains the CNV segments,
the column named 'clones' with the number of clones containing the event
and the column CCF with the total CCF of the event
}
