% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{getCombinedSNV}
\alias{getCombinedSNV}
\title{total SNV events}
\usage{
getCombinedSNV(X)
}
\arguments{
\item{X}{cnvobj}
}
\value{
a GRanges with all the events in the SNV list
}
\description{
Merges the snv list and sums the VAF (or CCF) of events found in multiple clones
In the end the result contains the GRanges of the events
a column called 'clones' with the number of clones that have this event
and the overall VAF of the event
}
