% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{getName}
\alias{getName}
\title{Get the name of the sample in a cloneobj}
\usage{
getName(X)
}
\arguments{
\item{X}{cloneobj}
}
\value{
the name of the object
}
\description{
The name is a string that can be used to identify the sample
}
