% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purity.R
\name{getPurity}
\alias{getPurity}
\title{Get the predicted purity for the sample}
\usage{
getPurity(X)
}
\arguments{
\item{X}{object of type cloneobj,}
}
\value{
the purity of the sample
}
\description{
Get the predicted purity for the sample
}
