% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishplot.R
\name{infer.clonal.models.f}
\alias{infer.clonal.models.f}
\title{infer the clonal model for a cloneobj}
\usage{
infer.clonal.models.f(
  freqs,
  founder = NULL,
  model = "monoclonal",
  type = "CCF"
)
}
\arguments{
\item{freqs}{A dataframe with the frequencies of mutations}

\item{founder}{the clone that is considered to be the founder. If set to NULL the founder will be the clone with the highest CCF or VAF}

\item{model}{the model of evolution to simulate, (monoclonal|polyclonal)}

\item{type}{type of frequencies provided (VAF|CCF)}
}
\description{
infer the clonal model for a cloneobj
}
