% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\name{inferSex}
\alias{inferSex}
\title{Infer the sex for a sample}
\usage{
inferSex(gr)
}
\arguments{
\item{X}{object of type GRanges or GRangeslist}
}
\value{
the sex of the sample as XX for female and XY for male
}
\description{
If events occur on the Y chromosome the sample is assigned to M
otherwise F
}
