% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CloneCNA.R
\name{loadCloneCNA}
\alias{loadCloneCNA}
\title{Generate a CNV object with the segments from facets and additional information}
\usage{
loadCloneCNA(clonecna, sample_name = NULL)
}
\arguments{
\item{clonecna}{FACETS object (as provided by facets) or a file in .rds format with a FACETS object. This object is typically saved as "fit" and contains "fit$cncf", "fit$purity" and "fit$ploidy"}

\item{sample_name}{name of the sample to be used in the object}
}
\description{
We keep all the events the program reports
}
