% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfreq.R
\name{loadSuperFreq}
\alias{loadSuperFreq}
\title{Create a cloneobj from SuperFreq}
\usage{
loadSuperFreq(method_dir, method_name = "SuperFreq", purity = NULL, ploidy = 2)
}
\arguments{
\item{method_dir}{directory created by superfreq. Under it are expected the directories data and rivers}

\item{method_name}{name of the method, defaults to SuperFreq}

\item{purity}{purity of sample, overrides the purity detected by the method, but does not change the CCF of clones}

\item{ploidy}{ploidy of sample, default is 2}
}
\description{
Create a cloneobj from SuperFreq
}
