% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCAT.R
\name{parseASCAT}
\alias{parseASCAT}
\title{Generate a cnv Granges from a dataframe created by ASCAT}
\usage{
parseASCAT(ascat.df)
}
\arguments{
\item{ascat.df}{input dataframe with CNV segments generated by ASCAT,}
}
\value{
A GRanges object with the segments predicted by ASCAT
}
\description{
Generate a cnv Granges from a dataframe created by ASCAT
}
