% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCAT.R
\name{parseASCAT_file}
\alias{parseASCAT_file}
\title{Generate a cnv Granges from a tab delimited created by ASCAT}
\usage{
parseASCAT_file(ascat.fn)
}
\arguments{
\item{ascat.fn}{input dataframe with CNV segments generated by ASCAT,}
}
\value{
A GRanges object with the segments predicted by ASCAT
}
\description{
Generate a cnv Granges from a tab delimited created by ASCAT
}
