% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCNV.R
\name{parseCfreec}
\alias{parseCfreec}
\title{Generate a cnv Granges from a dataframe created by ControlFreec (file ending in .CNVs or contains pvalues.)
Control-FREEC is a tool for detection of copy-number changes and allelic imbalances (including LOH)
using deep-sequencing data. It automatically computes, normalizes, segments copy number and beta allele
frequency (BAF) profiles, then calls copy number alterations and LOH. The control (matched normal) sample
is optional for whole genome sequencing data but mandatory for whole exome or targeted sequencing data.
Control-FREEC up to now (Jan 2021) is the standard tool used in the hCELG and hBMS pipelines for data
analysis of WES and WGS datasets.
When the genotype is - it means that the progrm did not find any SNPs to decide the correct genotype, and
the estimated copy number depends on the logRR
The output includes
* chromosome
* start position
* end position
* copy.number
* status
* genotype
* uncertainty of the assignment
* WilcoxonRankSumTestPvalue
* KolmogorovSmirnovPvalue
Note: at least for v 11.5 it is not possible to understand how subclones are coded in the output files.
As a result we only parse the main clone information (i.e no subclones). A related ticket has been opened on GitHub to help understand the C-FREEC outputs (https://github.com/BoevaLab/FREEC/issues/84)
|---------|----------|--------------|
|total cn | minor cn |	type         |
|---------|----------|--------------|
|0        |	0 	    | HOMDEL       |
|0        |	NA 	    | HOMDEL       |
|1        |	0 	    | HETDEL       |
|1        |	NA 	    | HETDEL       |
|2        |	0 	    | NEUTLOH      |
|2        |	1 	    | NEUT         |
|2        |	NA 	    | NEUT/Unknown |
|2+       |	0 	    | AMP-LOH      |
|2+       |	1+ 	    | AMP          |
|2+       |	NA 	    | AMP-[LOH?]   |
|---------|----------|--------------|}
\usage{
parseCfreec(controlfreec.df, filterLowUncertainty = TRUE, pvalThreshold = 0.01)
}
\arguments{
\item{controlfreec.df}{input filename with CNV segments,
filterLowUncerntainty if set to TRUE segments with uncertainty equal to -1 will be discarded
pvalThreshold  threshold of pvalues to filter if the pvalues are included in the file}
}
\value{
A GRanges object with the segments predicted by ControlFreec
}
\description{
Generate a cnv Granges from a dataframe created by ControlFreec (file ending in .CNVs or contains pvalues.)
Control-FREEC is a tool for detection of copy-number changes and allelic imbalances (including LOH)
using deep-sequencing data. It automatically computes, normalizes, segments copy number and beta allele
frequency (BAF) profiles, then calls copy number alterations and LOH. The control (matched normal) sample
is optional for whole genome sequencing data but mandatory for whole exome or targeted sequencing data.
Control-FREEC up to now (Jan 2021) is the standard tool used in the hCELG and hBMS pipelines for data
analysis of WES and WGS datasets.
When the genotype is - it means that the progrm did not find any SNPs to decide the correct genotype, and
the estimated copy number depends on the logRR
The output includes
* chromosome
* start position
* end position
* copy.number
* status
* genotype
* uncertainty of the assignment
* WilcoxonRankSumTestPvalue
* KolmogorovSmirnovPvalue
Note: at least for v 11.5 it is not possible to understand how subclones are coded in the output files.
As a result we only parse the main clone information (i.e no subclones). A related ticket has been opened on GitHub to help understand the C-FREEC outputs (https://github.com/BoevaLab/FREEC/issues/84)
|---------|----------|--------------|
|total cn | minor cn |	type         |
|---------|----------|--------------|
|0        |	0 	    | HOMDEL       |
|0        |	NA 	    | HOMDEL       |
|1        |	0 	    | HETDEL       |
|1        |	NA 	    | HETDEL       |
|2        |	0 	    | NEUTLOH      |
|2        |	1 	    | NEUT         |
|2        |	NA 	    | NEUT/Unknown |
|2+       |	0 	    | AMP-LOH      |
|2+       |	1+ 	    | AMP          |
|2+       |	NA 	    | AMP-[LOH?]   |
|---------|----------|--------------|
}
