% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACETS.R
\name{parseFACETS}
\alias{parseFACETS}
\title{Generate a cnv Granges from a dataframe created by FACETS}
\usage{
parseFACETS(facets.df)
}
\arguments{
\item{facets.df}{Dataframe with FACETS output}
}
\value{
A GRanges object with the segments predicted by FACETS
}
\description{
Depending on the combination of the lcn,tcn values it adds a flag under cnv_flag
describing the type of the copy number abberation.
}
\details{
FACETS is an Allele-specific copy number analysis (ASCN) tool and open-source software with a broad application
to whole genome, whole-exome, as well as targeted panel sequencing platforms.

It is a fully integrated stand-alone pipeline that includes sequencing BAM file post-processing, joint segmentation of total- and allele-specific read counts, and integer copy number calls corrected for tumor purity, ploidy and clonal heterogeneity.
FACETS provides information about subclones, this information is used for the visualization and comparisons.

Because the germline (diploid) segments are marked as 2 CNV for major clone and 1 Copy for minor (just a convention of the tool),
we add a column with the percentage of the minor clone, by subtracting the percentage of the major clone from the calculated purity.

The columns of the dataframe are:
* ‘chrom’ the chromosome to which the segment belongs;

* ‘seg’ the segment number;

* ‘num.mark’ the number of SNPs in the segment;

* ‘nhet’ the number of SNPs that are deemed heterozygous;

* ‘cnlr.median’ the median log-ratio of the segment;

* ‘mafR’ the log-odds-ratio summary for the segment;

* ‘segclust’ the segment cluster to which segment belongs;

* ‘cnlr.median.clust’ the median log-ratio of the segment cluster;

* ‘mafR.clust’ the log-odds-ratio summary for the segment cluster;

* ‘cf’ the cellular fraction of the segment;

* ‘tcn’ the total copy number of the segment;

* ‘lcn’ the minor copy number of the segment.

The column lcn.cf.em is added to capture the expected abundance of the minor clone.
Segments with cf.em == 1 represent a single clone. If the tcn.em is 2 and lcn.em is 1 this is the germline. In all other cases cf.em represents the cellular fraction of the segment
Segments with lcn.em=NA segments are tricky due to the sparsity of het SNPs we cant determine lcn unambiguously.
}
