% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACETS.R
\name{parseFACETS_file}
\alias{parseFACETS_file}
\title{Load a FACETS output from a tab delimited file}
\usage{
parseFACETS_file(facets.fn)
}
\arguments{
\item{facets.fn}{Path to the input file}
}
\value{
A GRanges object with the segments predicted by FACETS
}
\description{
This function loads file generated by FACETS (ref)
it expects to find a tab delimited file
}
\details{
The columns are:
\itemize{
\item ‘chrom’ the chromosome to which the segment belongs;
\item ‘seg’ the segment number;
\item ‘num.mark’ the number of SNPs in the segment;
\item ‘nhet’ the number of SNPs that are deemed heterozygous;
\item ‘cnlr.median’ the median log-ratio of the segment;
\item ‘mafR’ the log-odds-ratio summary for the segment;
\item ‘segclust’ the segment cluster to which segment belongs;
\item ‘cnlr.median.clust’ the median log-ratio of the segment cluster;
\item ‘mafR.clust’ the log-odds-ratio summary for the segment cluster;
\item ‘cf’ the cellular fraction of the segment;
\item ‘tcn’ the total copy number of the segment;
\item ‘lcn’ the minor copy number of the segment.
}
}
