% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfreq.R
\name{parseSuperFreq}
\alias{parseSuperFreq}
\title{Main function to parse the SNV segments and SNV events from  SuperFreq.}
\usage{
parseSuperFreq(method_dir)
}
\arguments{
\item{method_dir}{directory created by superfreq. Under it are expected the directories data and rivers}
}
\description{
Superfreq produces a variety of files, for the normal (if present) and tumor samples.
file data/CNAsegments{samplename}.tsv has a list of the CNV segments with their specific information
file rivers/{samplename}-river.tsv contains the list of the mutations & CNV segments assigned to clones
file data/clones_{samplename}.tsv contains the list of the clones with their clonality
}
\details{
Our main file where we start all analyiss from is the rivers/{samplename}-river.tsv
and we use it as the scaffold, to add more information from the CNA segments and variants files.
main function for Superfreq
Create necessary objects from the superfreq directory
}
