% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfreq.R
\name{parseSuperFreq_CNV}
\alias{parseSuperFreq_CNV}
\title{Parse the CNV segments from SuperFreq}
\usage{
parseSuperFreq_CNV(method_dir)
}
\arguments{
\item{method_dir}{directory created by superfreq. Under it are expected the directories data and rivers}
}
\value{
GRanges object with all the CNV segments
}
\description{
Parse the CNV segments from SuperFreq
}
