% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superfreq.R
\name{parseSuperFreq_SNV}
\alias{parseSuperFreq_SNV}
\title{Parse the SNV events from SuperFreq}
\usage{
parseSuperFreq_SNV(method_dir)
}
\arguments{
\item{method_dir}{directory created by superfreq. Under it are expected the directories data and rivers}
}
\value{
GRanges object with all the CNV segments
}
\description{
Parse the SNV events from SuperFreq
}
