% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cloneobj}
\alias{plot.cloneobj}
\title{plot the clone information}
\usage{
\method{plot}{cloneobj}(X, overlay = FALSE, chromosomes = NULL)
}
\arguments{
\item{X}{cloneobj}
}
\description{
plot the clone information
}
