% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{plotAlluvialCNV}
\alias{plotAlluvialCNV}
\title{Internal method to plot an alluvial with distribution of events}
\usage{
plotAlluvialCNV(query, subject, wiggle = 0.1)
}
\arguments{
\item{query}{query the query cloneobj}

\item{subject}{the subject cloneobj}
}
\value{
a figure with distribution of events
}
\description{
Internal method to plot an alluvial with distribution of events
}
