% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCompare}
\alias{plotCompare}
\title{plot an integrated plot with CNV information from multiple samples}
\usage{
plotCompare(X, chromosomes = NULL, filter.neutral = TRUE, ncols = 5)
}
\arguments{
\item{X}{is a list of cnvobj is provided it is assumed that they will be compared.}

\item{chromosomes}{list of chromosomes to plot}

\item{filter.neutral}{filter regions that are NEUT cnv}

\item{ncols}{the number of columns to arrange the chromosome plots in a table format.}
}
\description{
Each plot is per chromosome
This plot contains at least four tracks
top: an ideogram of the chromosome
second from top: the location on the chromosome with coordinates
the following plots have the CNV of each segment
and the CCF for each segment
}
\details{
If this function is provided with a reference cnvobj
an additional plot with the reference CNVs is added below

If this function is used to compare two cnvobj objects
each panel has overlapping histograms with different colors
}
