% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDensity}
\alias{plotDensity}
\title{get the CCF of a subclone}
\usage{
plotDensity(X)
}
\arguments{
\item{X}{cloneobj ,}
}
\value{
an plot object from ggplot
}
\description{
The track contains the segments and the CCF of each of the segments
plot the density of the CCF for an object
}
\details{
Depending if SNVs or CNVs are present in the object one can get
a density plot with a single line or two lines.
}
