% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{print.cloneobj}
\alias{print.cloneobj}
\title{Print information for the object}
\usage{
\method{print}{cloneobj}(x, ...)
}
\description{
Provide a table with the clonal information of the object
}
\details{
The table contains:

known clones

purity of the sample

ploidy of the sample
}
