% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{seqlevelsStyle<-}
\alias{seqlevelsStyle<-}
\title{Override the GenomeInfoDb seqlevelstyle for cloneobj}
\usage{
seqlevelsStyle(X) <- value
}
\arguments{
\item{X}{the cloneobj}

\item{style}{specify the style}
}
\description{
Override the GenomeInfoDb seqlevelstyle for cloneobj
}
