% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{seqlevelsStyle}
\alias{seqlevelsStyle}
\title{Override the GenomeInfoDb seqlevelstyle for cloneobj}
\usage{
seqlevelsStyle(X)
}
\arguments{
\item{X}{the cloneobj}
}
\value{
current style in the object
}
\description{
Override the GenomeInfoDb seqlevelstyle for cloneobj
}
