% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{setName}
\alias{setName}
\title{Set the name of the sample in a cloneobj}
\usage{
setName(X, sample_name)
}
\arguments{
\item{X}{cloneobj}

\item{sample_name}{the name of the object}
}
\value{
an updated cloneobj
}
\description{
The name is a string that can be used to identify the sample
}
