% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{setPloidy}
\alias{setPloidy}
\title{Set the predicted ploidy for the sample}
\usage{
setPloidy(X, v)
}
\arguments{
\item{X}{object of type cloneobj,}

\item{v}{ploidy (float)}
}
\value{
the updated object X with the ploidy set
}
\description{
Two parameters are set:
}
\details{
**ploidy** is the rounded ploidy (e.g. 2.1 will become 2)

**real_ploidy** is the actual value provided by the user
}
