% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purity.R
\name{setPurity}
\alias{setPurity}
\title{Set the predicted purity for the sample}
\usage{
setPurity(X, v)
}
\arguments{
\item{X}{object of type cloneobj,}

\item{v}{purity (0-1)}
}
\value{
the updated object X with the purity set
}
\description{
Set the predicted purity for the sample
}
