% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\name{setSex}
\alias{setSex}
\title{Set the sex for the sample}
\usage{
setSex(X, v)
}
\arguments{
\item{X}{object of type cloneobj,}

\item{v}{purity (0-1)}
}
\value{
the updated object X with the purity set
}
\description{
Set the sex for the sample
}
