% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneobj.R
\name{cloneobj}
\alias{cloneobj}
\title{create the object}
\usage{
cloneobj(
  purity = NA,
  ploidy = NA,
  cnvlist = NULL,
  snvlist = NULL,
  method = NULL,
  genome = "hg38"
)
}
\arguments{
\item{purity}{(float) Purity of the sample ,}

\item{ploidy}{(float) Ploidy of the sample,}

\item{cnvlist}{GRangesList with cnv clones,}

\item{snvlist}{GRangesList with snv clones,}

\item{method}{(list of string) Name of method used to predict these values}

\item{genome}{the name of the genome (hg19, hg38)}
}
\value{
an object of type cloneobj
}
\description{
cloneobj is a structure that holds information
for the clones
}
\details{
This information includes

**Method name** the name(s) of the method(s) used to create the data for this object

**Purity**: the predicted purity of the sample (0-1)

**Ploidy**: the predicted ploidy of the sample (default 2)

**Copy number clones** GRangesList with segments from one or multiple clones (depending on the tool used)

**Sample Sex** either provided by the user or inferred

**SNV clones** GRangesList of clones defined by the SNV VAF (or CCF)
}
