% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishplot.R
\name{fishPlot}
\alias{fishPlot}
\title{create a fishplot using the snv events (if available)}
\usage{
fishPlot(x, sname = "sample")
}
\arguments{
\item{x}{a cloneobj}

\item{sname}{name of sample to show on the plot}
}
\description{
create a fishplot using the snv events (if available)
}
