% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{getPloidy}
\alias{getPloidy}
\title{Get the predicted ploidy for the sample}
\usage{
getPloidy(X)
}
\arguments{
\item{X}{object of type cloneobj,}
}
\value{
a list with the _real.ploidy_ and _ploidy_
}
\description{
Two parameters are returned:
}
\details{
**ploidy** is the rounded ploidy (e.g. 2.1 will become 2)

**real_ploidy** is the actual value provided by the user
}
