% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACETS.R
\name{loadFACETS}
\alias{loadFACETS}
\title{Generate a CNV object with the segments from facets and additional information}
\usage{
loadFACETS(facets, maxclones = 4)
}
\arguments{
\item{facets}{FACETS object (as provided by facets)}
}
\description{
We keep only the non NEUTRAL events
}
\details{
We store each clone (as defined by the ccf) separately in the object.
}
